Move useTestRunner file to its own directory
