+++ src/NetworkTestWidget/NetworkTestWidget.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Button, Typography } from '@material-ui/core';
-import useTestRunner from './useTestRunner';
+import useTestRunner from './useTestRunner/useTestRunner';
 import RegionResult from './RegionResult/RegionResult';
 import { Region } from '../types';
